requirements.txt
dspy_ai==2.4.9
wikipedia==1.4.0
sentence-transformers>=2.2.2
langchain-text-splitters>=0.0.1
trafilatura>=1.6.1
langchain-huggingface>=0.0.2
qdrant-client>=1.7.0
langchain-qdrant>=0.0.1
numpy==1.26.4
litellm==1.59.3
diskcache>=5.6.3
requests>=2.31.0
backoff>=2.2.1
ujson>=5.8.0
pathlib>=1.0.1
tqdm>=4.66.1
httpx>=0.27.0
tavily-python>=0.5.1
boto3>=1.34.0
anthropic>=0.18.1
openai>=1.12.0
transformers>=4.38.0